🎨style: 검증 파일 이름 수정
- Validator라는 함수 하나로 관리하면 중요도가 높아질 것 같아서 분리해서 관리하기로 했습니다.
